64bit compiler fixes

+++ lemon.c
@@ -1563,7 +1563,8 @@ char **argv)
 /*
 ** Return a pointer to the next structure in the linked list.
 */
-#define NEXT(A) (*(char**)(((unsigned long)A)+offset))
+//#define NEXT(A) (*(char**)(((unsigned long)A)+offset))
+#define NEXT(A) (*(char**)((A)+offset))
 
 /*
 ** Inputs:
